Move TargetData to its own file
