+++ Assets/Scripts/Targets/TargetData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e7dc87ec56a6cb0469a009674fc3f3da
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
